refactor(backend): modification of use cases and views
The use cases were refactored in such a way that the logic of future functionalities for each type of user can be implemented more easily. The existing views were also refactored, now they will be able to manage any type of request in a more robust and orderly manner based on the HTTP method. The rest of the code was adjusted.